+++ packages/frontend/src/components/dialogs/CreateChat/index.tsx
@@ -17,6 +17,7 @@ import {
   PseudoListItem,
   PseudoListItemAddMember,
   PseudoListItemAddContact,
+  PseudoListItemAddContactOrGroupFromInviteLink,
 } from '../../helpers/PseudoListItem'
 import GroupImage from '../../GroupImage'
 import { runtime } from '@deltachat-desktop/runtime-interface'
@@ -57,6 +58,7 @@ import { ContextMenuContext } from '../../../contexts/ContextMenuContext'
 import ImageCropper from '../../ImageCropper'
 import { RovingTabindexProvider } from '../../../contexts/RovingTabindex'
 import ViewProfile from '../ViewProfile'
+import { isInviteLink } from '@deltachat-desktop/shared/util'
 
 type ViewMode = 'main' | 'createGroup' | 'createBroadcastList'
 
@@ -162,8 +164,13 @@ function CreateChatMain(props: CreateChatMainProps) {
       contactCache[contactIds[0]]?.address.toLowerCase() ===
         queryStr.trim().toLowerCase())
   )
+  const showPseudoListItemAddContactFromInviteLink =
+    queryStr && isInviteLink(queryStr)
   const contactsAndExtraItems = useMemo(
     () => [
+      ...(showPseudoListItemAddContactFromInviteLink
+        ? [CreateChatExtraItemType.INVITE_LINK]
+        : []),
       ...(needToRenderAddContactQRScan
         ? [CreateChatExtraItemType.ADD_CONTACT_QR_SCAN]
         : []),
@@ -180,6 +187,7 @@ function CreateChatMain(props: CreateChatMainProps) {
       needToRenderAddContact,
       needToRenderAddContactQRScan,
       needToRenderAddGroup,
+      showPseudoListItemAddContactFromInviteLink,
     ]
   )
 
@@ -389,6 +397,7 @@ function CreateChatMainRow({
   const item = contactsAndExtraItems[index]
 
   const tx = useTranslationFunction()
+  const accountId = selectedAccountId()
 
   const el = (() => {
     switch (item) {
@@ -432,6 +441,14 @@ function CreateChatMainRow({
           />
         )
       }
+      case CreateChatExtraItemType.INVITE_LINK: {
+        return (
+          <PseudoListItemAddContactOrGroupFromInviteLink
+            inviteLink={queryStr!}
+            accountId={accountId}
+          />
+        )
+      }
       default: {
         const contact: Type.Contact | undefined = contactCache[item]
         if (!contact) {
@@ -466,6 +483,7 @@ const enum CreateChatExtraItemType {
   ADD_GROUP,
   ADD_BROADCAST_LIST,
   ADD_CONTACT,
+  INVITE_LINK,
 }
 
 type CreateGroupProps = {