+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - tauri: support for webxdc #4740, #4852
 - create chat: add context menu option to view profile #4880
+- create chat: add support for invite links to search bar
 
 ### Changed
 - switch to account the webxdc is from when sending to chat (tauri and electron edition) #4740